More notes about difference between VICE and actual hardware behavior
VICE seems to think the CPU can change a register and have it
take effect on the 2nd dot of its own phase which is unlikely
in real hardware.  It's most likely the case that actual
hardware sees the register change after the falling edge of phi.
Keeping like this for now.