Fix issue with sprite pixel reads
Previous change (to use cycles) ended up vausing Vivado
to infer latches for ba_sprite which appears to have caused
timing issues.  Taking a different approach where we have
a separate sprite_raster_x register that is offset so that
it resets to 0 when the first sprite BA low should happen.
Then everything is relative to 0 and we can use a simple
comparison interval like we did with xpos but don't have
to worry about wrap around conditions or repeats (R8).

This seems to have cleaned up sprite accesses.  Although
there are still some glitches at the top of MsPacMan.