Get rid of start_of_* flags
This was causing grief for irq handling.  It was done to
match VICE raster_line increments but I don't think the real
device works this way.  raster_line increment in VICE is delayed
by 1/2 cycle on cycle 0 after each line starts and then delayed again
on frame start.  But this is likely just a VICE thing and doesn't
make sense for the fpga.  It was causing irq to get raised again
at cycle 0 because raster line was not yet incremented.