Fix char/color buf on first sync and sprite collision
We were not syncing char/color buf properly on sync so
any frame we tried to sync to that suppressed badlines
would never read in char/colors and we'd just get black
for all pixels.  This syncs the internal buffer from
VICE's buffer so we can sync mid-frame without issue.

Moved timing of sprite-sprite and sprite-data so that
sprite5 and sprite6 tests are consistent in their
warnings. Documented the warning but may not be a problem.

This gets lunatico massively interleaved sprite crunch
working (on sim) with no major warnings except 0x1f reg
on border area. Need to investigate.