Fix two issues with Uncensored ski hill
It appears that a higher priority sprite's desire to not
overwrite pixels takes precedence over a lower priority
sprite's desire to overwrite a fg pixel.  Need more
testing but this appears to have been the problem with
Uncensored skil hill color mismatches on the sides.

Also stole VICE's magic delayed bmm for use in address
generation.  Seems we are supposed to delay bmm falling
but not rising?  Weird.