Clock stuff
Using one MMCM for two clocks for on-board
clocks for both cmod and qmtech.  Not worth
extra complexity for more accurate clocks
for something that will never be used.  On
board clocking gets close enough to required
color clocks for an image but colors will
shimmer.  Just use as a fallback if no
external clock is available.