Fix sprite collisions on RHS border pixel
It appears that even though a pixel is rendered as a border
pixel on the RHS, sprite to data collisions still take
place there.  Lunatico MISC will fail without the
extra condition placed in this CL to consider a low to high
transition as capable of triggering sprite collisions while
still being rendered as a border pixel.