Rewrite pixel sequencer
Rewriting the pixel sequencer to be pixel
perfect match to VICE. Added logging to
match logging in VICE so we can investigate
differences in output more easily. Use the same
pixel value scheme as VICE and add logging that
matches VICE hooks so we can perform diffs that
will point to issues.