Fix border issue
Borders were not flipping at the right time
Instead of changing xpos on the comparison, making this
equal to VICE but then delaying the outputs before they
get into the pixel sequencer.

Also, can get rid of the right border hack in sprites. This
was the cause.

Fixes test4.prg and vborder2-36.prg