Going back to dot4x from sysclock
Back to one clock output from CMT
But using another one to generate the dot8x
This would not be necessary on real hardware
with a proper color clock.  But we have a 50mhz
clock with too little accuracy and the difference
in dot4x causes crashes.  Generating dot8x
from dot4x requires a special constriant flag
to allow this as it is considered bad design
practice.