Fix sprite to data collision issue
The border should not influence whether
sprite to data collisions happen.  If CSEL is
off, sprite to data collisions should still happen
underneath the extended border.  This was breaking
lunatico.  There should be no need to check border
because it should be impossible for pixels to be
set outside the visible range.