Use is_15khz flag
Modify sync module to use the is_15khz flag
and use native resolution for height (still
doubled for width).  Might work for monitors
that support 15khz H freq but is untested.
Also sync module should not be setting
colors outside of active, this is done by
final color index lookup so there was a
black stripe on left side of the screen.