Register reads should return last desired values
The register reads on 15khz and chip
model should return what the last persisted values
were, not what is presenting active. Since we
apply raster lines immediately, these are always
one and the same but the others should not
change while the system is running.  This was
confusing the config utility that was trying
to change things but then would read the old
value and thought nothing changed.  Consider
splitting the bytes for these flags so that
the upper nibble is for active value and
the lower nibble is for last persisted but
not yet active value.