Bump colors from 12-bit to 18-bit
This CL increases the color space to
6 bits per RGB channel.  This will
make any software that sets the palette
look dark since prior to this cl only
the lowest 4 bits were considered.
Changes are still necessary to update
the simulator.  Smoke test of
uncensored passed and ntsc/pal switch
was okay.

The 'new' upper bits have been routed
to unused pins on the mojov3/hat combo
and are not accessible unless the pins
are either bent or some wires soldered
on the top somehow.  But HDMI and
VGA are still functional, just with
the lowest 2 bits not used.