Fix setup timing violation in comp_sync
Need to handle clock domain crossing for signals
used in the clk_col16 process block.  This fixes
the setup timing problems reported in ISE when
configurable lumas is not enabled.  There are still
problems with configurable lumas enabled but this
likely won't be in the final release and is only
for fine tuning so not going to fix it just yet.

Also one of the amplitude registers had a range
off by 1.