Plumb tx_busy and rx_busy through to a register bit
This CL passes through the tx_busy | rx_busy combo
as bit 5 of the vmem flags register so the 6502 can
safely poke persisted register changes without
overflowing the serial buffer. It's rare for tx_busy
to be high since the transmission of the serial data
happens much faster than the 6502 can possibly
poke.  However, rx_busy is more common since it
takes some time for the MCU to write to its EEPROM.

Next CL to update loader is required for this to
work.