Add native resolution support to sim and fix v/h sync
When no rgb is available, use the pixel_color3 value
to lookup colors. This is the case when composite only is
available.

Also switching composite sync module to use raster_x rather
than xpos and fixing the hsync/vsync ranges like we did in
the vga sync module.  The durations after hsync start are
fixed but its location can be adjusted with the very first
offset after hvisible start. (Need to make it configurable)