Make config utils init eeprom
Config utils should be able to detect when
the eeprom is not initialize and give the
option to init before entering the main
menu.  Will init with good known defaults
before proceeding.

Also changed MCU loader to use slower
but compatible flash method of byte by byte.
The replacement chip is not compatible with
the faster method.

Untested