Fix eeprom behavior
Use full 1024 bytes of EEPROM to save 4 sets of register values
for each of the chips we emulate.  EEPROM read and write
changed to use an eeprom_bank register which selects the
bank.  The bank will be set to the chip at startup.  Bank
0 is always read first until chip is read. Then the bank
switches to that chip.  So all registers below chip will
be cross-chip registers while everything above is per-chip.
Had to move colors to 0x40.  Also moved magic bytes to 0x00
to avoid the weird offset at the start.