Maybe fix EEPROM read at boot issue
Last flash was not reading magic on 2nd
iteration after reset.  Now it does but
not entirely sure if this fixes it.  Some
flash images seem to read bad magic
bytes (consistently the same) and I don't
know why.  With this change, the data
is consistently read okay but that might
just be because the bitstream changed.