Fix EEPROM / FLASH bug?
This *may* fix the eeprom/flash bug. I suspect
the setup time for the spi_d line was not being
met. This delays the clock so that we prep the
data line about 30ns before the clock edge of
either the flash chip clock or eeprom chip clock
signals.

Also fixed versioning problem with Makefile.inc
Was copying over version we set with perl.