Fix sprite issues on Efinix Trion
It looks like having a blocking assignment in
one process block and having another process block
'pick up' that value is not well defined in verilog.
On Xilinx, the delay pipeline we had would
consistently work as expected (and in verilator too)
but on Efinix it would sometimes 'see' the result
of the blocking assignment a whole pixel early,
resulting in 'missing' pixels in the pixel sequencer.
Moving the delay pipeline into the same process block
that makes the blocking assignments (i.e.
sprite_cur_pixel1 and active_sprite) fixed the issue.