Fix problem with golden images for efinix
Page size has to be 4k because the efinix places the 2nd bitstream
immediately after the first and it isn't divisible by 16k, so the
last segment would always fail verification because it would bleed
into the next (multiboot) image bytes.

Also fixed an issue with hex to bit converter not handling lower
case which was making golden bitstreams not load.