Use a slightly slower dot clock for DVI
Change timing for DVI output on efinix to get a more
stable/higher quality DVI output.  Use a different
sync/pixel generator for Efinix/DVI so we can use a
different pixel clock than that of the 4xdot we use for
everything else. This has the benefit of chopping off
some of the excessive border we get with the full
dot clock rate. Down side is the analog RGB will
need to be a separate core.