Try to improve DVI sync
Noticed DVI timings were out of spec.  DVI spec says
hsync must be at least 128 pixels wide.  Our sync period
was shy for all 3 models.  Bringing sync period up to 128
(64 for native res because is doubled in x dir).  Also
made sure that serializer was actually using the reset
signal.  So now the clock and control registers don't
get activated until after the chip is selected.

This seems to have improved NTSC syncing a lot on this
particular monitor and a bit of improvement for PAL.
PAL failure to sync on LG monitor still needs
investigating but prior to this CL, the LG monitor
was not syncing at all or had frequent black
screen boots.