Add checksum to flash program
Add checksum bytes to the end of every segment.
Drive integrity check will try the first few
segments and ensure the checksums match.  If
not, flash will be aborted before it is
erased.

Had to move scratch space from $5000 to $6000
and the loader from $9000 to $a004.  Spartan
flash images unlikely to be supported but keeping
the 16k space available just in case.